requirements/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/requirements.in
#
amqp==2.5.2               # via kombu
asgiref==3.2.3            # via django
attrs==19.3.0             # via -r requirements/requirements.in
billiard==3.6.1.0         # via celery
celery-progress==0.0.9    # via -r requirements/requirements.in
celery[redis]==4.4.0      # via -r requirements/requirements.in
certifi==2019.11.28       # via requests
chardet==3.0.4            # via requests
coreapi==2.3.3            # via -r requirements/requirements.in
coreschema==0.0.4         # via coreapi
defusedxml==0.6.0         # via python3-openid
django-allauth==0.41.0    # via -r requirements/requirements.in
django-anymail[mailgun]==7.0.0  # via -r requirements/requirements.in
django==3.0.5             # via -r requirements/requirements.in, django-allauth, django-anymail, djangorestframework
djangorestframework==3.11.0  # via -r requirements/requirements.in
idna==2.8                 # via requests
itypes==1.1.0             # via coreapi
jinja2==2.10.3            # via coreschema
kombu==4.6.7              # via celery
markupsafe==1.1.1         # via jinja2
oauthlib==3.1.0           # via requests-oauthlib
psycopg2-binary==2.8.3    # via -r requirements/requirements.in
python3-openid==3.1.0     # via django-allauth
pytz==2019.3              # via celery, django
redis==3.3.11             # via celery
requests-oauthlib==1.3.0  # via django-allauth
requests==2.22.0          # via coreapi, django-allauth, django-anymail, requests-oauthlib
six==1.13.0               # via django-anymail
sqlparse==0.3.0           # via django
uritemplate==3.0.1        # via coreapi
urllib3==1.25.7           # via requests
vine==1.3.0               # via amqp, celery
